Fix: friends_book Container clé doit être long (pas int)
IRecord.Id est typé long, le ORM appelle ConcurrentDictionary.Add(object,
object) avec un long key. Avec un dict <int, ...> ça crash avec
ArgumentException dès le 1er .AddLater().

+++ server/giny/Sources/Servers/Giny.World/Records/Social/OneAirFriendsBookRecord.cs
@@ -13,8 +13,10 @@ namespace Giny.World.Records.Social
     [Table("friends_book")]
     public class OneAirFriendsBookRecord : IRecord
     {
+        // ORM Giny utilise IRecord.Id (long) comme clef du Container, il faut
+        // donc un dict long-keyed même si la colonne SQL est INT.
         [Container]
-        private static readonly ConcurrentDictionary<int, OneAirFriendsBookRecord> Books = new ConcurrentDictionary<int, OneAirFriendsBookRecord>();
+        private static readonly ConcurrentDictionary<long, OneAirFriendsBookRecord> Books = new ConcurrentDictionary<long, OneAirFriendsBookRecord>();
 
         [Ignore]
         public long Id => AccountId;
@@ -41,7 +43,7 @@ namespace Giny.World.Records.Social
 
         public static OneAirFriendsBookRecord Get(int accountId)
         {
-            return Books.TryGetValue(accountId, out var book) ? book : null;
+            return Books.TryGetValue((long)accountId, out var book) ? book : null;
         }
 
         public static OneAirFriendsBookRecord GetOrCreate(int accountId)
